modification bar changes / implementation
\backend\api\entriesmod.js
     -> apiPostEntry now takes in an HID input
          apiPostEntry now inputs with an HID input
\backend\dao\modDAO.js
     -> addEntry now takes in an HID input
          updateEntry now takes in an HID input, locates the database object to update using HID instead of TITLE.
\src\pages\entry.js
          entry location from database now determines whether or not there is an HID prop passed in through another access by checking for "hid", not "id".
\srs\pages\entryMod.js
     -> Added basic implementation for deleteEntry, updateEntry, and postEntry. deleteEntry now takes in an _ID input from the text field and checks if it has a length of 24. If it does, it sends a delete request through that _ID. If not, it determines that the deletion ID input is not valid.
updateEntry now uses the input string to generate an "updateObject", which is, if it contains an hid property to use for locating the updated object, which is then modified (overwritten) by the update request object and sent back to the database for updating.
postEntry now uses the input string to generate an "addObject", which, if it has values for the required properties (title, source, keywords, url, country), is given a generated HID. HID generation requires retrieving the total amount of other objects which already have the given ISO-country and Book-Code combination; the added entry will be the Nth object of that combination, starting at 1. 0s are added onto the Nth-object value to create a properly-formatted HID, which is then added to the "addObject", and an object is finally added to the database containing te addObject's contents.
     -> Fixed subString check for verifying a post command; check up to 4 only checks the first 4 characters of the string (post), when we need to actually check the first 5 characters of the string (post:)
src\dev\countryList.js
     ->Moved the pair for UNKNOWN country (UNK:"UN") to the TOP of te countryList in format.
     -> Added regional properties for Senegambia (SGM), Abyssinia (AB), and Gold Coast (GC).
src\dev\idfromName
     -> Added an "idfromName" object which contains pairs giving individual Book-names corresponding ID codes.